# internal/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: EEBUS Testbench
  description: >
    Reference-peer control plane for testing an EEBUS device under test. Every panel in
    the dashboard at /ui is a thin wrapper over this REST API -- anything the UI can do,
    a script can do too (see examples/*.py for standalone httpx clients with no facade
    imports).
  version: 0.1.0-go
...
paths:
...
  /api/v1/config/reload:
    post:
      tags: [config]
      summary: Re-read config/eebus.yaml and push the new peer list to the running stack (static mode only)
      responses:
        "200": { description: reloaded }
        "400": { description: invalid config, not reloaded }
  /api/v1/config/active-stack:
    put:
      tags: [config]
      summary: Start the named counterparty (there is only one -- eebus-go-remote)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                stack: { type: string, example: eebus-go-remote }
                persist: { type: boolean, default: false }
      responses:
        "200": { description: OK }
        "404": { description: unknown stack }
        "501": { description: "persist:true requested but config write-back isn't built yet" }
...
  /api/v1/stacks/{id}:
    get:
      tags: [stacks]
      summary: One stack's status
      parameters: [{ $ref: "#/components/parameters/StackId" }]
      responses: { "200": { description: OK }, "404": { description: unknown stack } }
  /api/v1/stacks/{id}/start:
    post:
      tags: [stacks]
      summary: (Re-)start the embedded EEBUS stack
      parameters: [{ $ref: "#/components/parameters/StackId" }]
      responses: { "200": { description: OK }, "404": { description: unknown stack } }
  /api/v1/stacks/{id}/stop:
    post:
      tags: [stacks]
      summary: Stop the embedded EEBUS stack (SHIP/SPINE only -- the dashboard HTTP API stays up)
      parameters: [{ $ref: "#/components/parameters/StackId" }]
      responses: { "200": { description: OK }, "404": { description: unknown stack } }
  /api/v1/stacks/{id}/logs:
    get:
      tags: [stacks]
      summary: Tail this process's own console log (in-memory ring buffer, not a file)
      parameters:
        - { $ref: "#/components/parameters/StackId" }
        - name: tail
          in: query
          schema: { type: integer, default: 200 }
      responses: { "200": { description: OK } }
...
  /api/v1/peers/{ski}/usecases:
    get:
      tags: [peers]
      summary: Use cases this peer itself advertises
      parameters: [{ $ref: "#/components/parameters/Ski" }]
      responses: { "200": { description: OK } }
  /api/v1/peers/{ski}/profile:
    get:
      tags: [peers]
      summary: Connected state, entity types, and advertised use cases for one peer
      parameters: [{ $ref: "#/components/parameters/Ski" }]
      responses: { "200": { description: OK }, "404": { description: not connected } }
  /api/v1/peers/{ski}/manufacturer:
    get:
      tags: [peers]
      summary: Device classification (manufacturer data) of every entity of the peer, read from the device
      parameters: [{ $ref: "#/components/parameters/Ski" }]
      responses: { "200": { description: OK }, "404": { description: not connected } }
  /api/v1/peers/{ski}/subscriptions:
    get:
      tags: [peers]
      summary: Subscriptions the peer holds on our features, and per feature type of ours the types of the peer's entities subscribed to it (a vehicle that implements OPEV scenarios 2 and 3 appears under DeviceDiagnosis)
      parameters: [{ $ref: "#/components/parameters/Ski" }]
      responses: { "200": { description: OK }, "404": { description: not connected } }
  /api/v1/peers/{ski}/trust:
    post:
      tags: [peers]
      summary: Trust (connect to) a peer -- also how you approve a pending pairing request
      parameters: [{ $ref: "#/components/parameters/Ski" }]
      responses: { "200": { description: OK } }
    delete:
      tags: [peers]
      summary: Untrust and disconnect a peer
      parameters: [{ $ref: "#/components/parameters/Ski" }]
      responses: { "200": { description: OK } }
  /api/v1/peers/{ski}/deny:
    post:
      tags: [peers]
      summary: Deny an incoming pairing request
      parameters: [{ $ref: "#/components/parameters/Ski" }]
      responses: { "200": { description: OK } }
  /api/v1/discover:
    post:
      tags: [discover]
      summary: Real zeroconf mDNS scan for _ship._tcp devices on the network (read-only, a few seconds)
      parameters:
        - name: timeout_s
          in: query
          schema: { type: number, default: 5 }
      responses: { "200": { description: OK } }
...
  /api/v1/energy/{ski}/snapshot:
    get:
      tags: [energy]
      summary: Aggregated live snapshot across LPC/LPP/MPC/MGCP/EV/PV/battery for one peer
      parameters: [{ $ref: "#/components/parameters/Ski" }]
      responses: { "200": { description: OK }, "404": { description: not connected } }
  /api/v1/energy/{ski}/history:
    get:
      tags: [energy]
      summary: Recorded snapshot history (in-memory, up to 12h)
      parameters:
        - { $ref: "#/components/parameters/Ski" }
        - name: seconds
          in: query
          schema: { type: number, default: 3600 }
      responses: { "200": { description: OK } }
    delete:
      tags: [energy]
      summary: Clear recorded history for one peer
      parameters: [{ $ref: "#/components/parameters/Ski" }]
      responses: { "200": { description: OK } }
  /api/v1/events/recent:
    get:
      tags: [events]
      summary: Ring buffer of recent lifecycle/usecase/spine events
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 200 }
        - name: level
          in: query
          schema: { type: string, enum: [lifecycle, usecase, spine] }
        - name: ski
          in: query
          schema: { type: string }
      responses: { "200": { description: OK } }
...
  /api/v1/trace:
    get:
      tags: [trace]
      summary: Raw SHIP frames as captured on the wire (pre-repair), with conformance findings; list responses omit the raw payload
      parameters:
        - name: after
          in: query
          schema: { type: integer }
          description: Return only frames with a sequence number greater than this (cursor polling)
        - name: limit
          in: query
          schema: { type: integer }
        - name: ski
          in: query
          schema: { type: string }
        - name: dir
          in: query
          schema: { type: string, enum: [send, recv] }
        - name: findings
          in: query
          schema: { type: string, enum: [only] }
          description: Pass "only" to return exclusively non-conformant frames
      responses: { "200": { description: OK } }
    delete:
      tags: [trace]
      summary: Drop the retained frames and reset the per-peer conformance sessions
      responses: { "200": { description: OK } }
...
  /api/v1/trace/{seq}:
    get:
      tags: [trace]
      summary: One captured frame in full, including the raw wire payload
      parameters:
        - name: seq
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200": { description: OK }
        "404": { description: Frame no longer retained }
...
  /api/v1/scenarios/{name}/run:
    post:
      tags: [scenarios]
      summary: Run one scenario by ID
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      responses: { "200": { description: OK }, "404": { description: no such scenario } }
...
components:
  parameters:
...
    Ski:
      name: ski
      in: path
      required: true
      schema: { type: string }
      description: 40 hex character SKI
    StackId:
      name: id
      in: path
      required: true
      schema: { type: string, example: eebus-go-remote }
...